intelligence-engine/requirements.txt
mysql-connector-python
pdfplumber
pymupdf
Pillow
podcastfy
fastapi
uvicorn[standard]
python-multipart